Interpret the results in the center of the screen visible
Since we do not observe the top and bottom of the screen with our eyes,
we may not notice the results there, so interpret only the results in
the center of the screen as visible.

Since this could be the case if you want to jump to the same result
twice (for example, if the result is at the very bottom of the page), I
refined the algorithm to continue searching at the other end of the
page.