Increase comfort
The toolbar buttons have a larger press surface for easy use with
fat-finger.

The buttons in the sidebar only appear if there is a search term, so it
is a little clearer what they are for. The buttons appear above the
search box so that they are not obscured by the popup list.